refactor : 패키지 형식 재정리
타 프로젝트 진행 과정에서 배운대로 정리해보기 ( 도메인 별 분류 및 패키지 명 변경 )
